requirements.txt
torch
torchvision
torchaudio
numpy
scikit-learn
matplotlib
seaborn
opencv-python
pillow
pydicom
grad-cam
streamlit
tqdm
